requirements.txt
boto3==1.11.13
botocore==1.14.13
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
future==0.18.2
idna==2.8
jmespath==0.9.4
motor==2.1.0
pkg-resources==0.0.0
pycparser==2.19
pymongo==3.10.1
python-dateutil==2.8.1
python-telegram-bot==12.3.0
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
